Split noteserver services into packages.
Add a plugin interface to the notifier package, and make each of the services a
package that implements the interface. The noteserver binary is now just a
patch to hook together the plugins and run the service. More methods and more
services can be added by adding additional package imports.