keygen: Enable a strict flag in the request message.
With this flag set, the server will report an error if the request mentions a
site that doesn't match one known to the configuration. This allows the user to
detect typographical errors. The userkey program sets the flag by default, but
you can override it using the -lax flag on the command line.