notes: Add support for other file extensions.
The default file extension is still .txt, but if the tag contains a period it
will now be used to specify an alternative, e.g., tag.md means to use a
filename like tag-19990101.md instead. When reading an existing file, a match
on a tag will accept any extension.